rest_api/pyproject.toml
[project]
name = "rest_api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12.11,<3.13"
dependencies = [
    "alembic==1.13.1",
    "apispec==6.3.1",
    "auth-client",
    "boto3==1.34.7",
    "clickhouse-connect==0.8.15",
    "cloud-adapter",
    "config-client",
    "cryptocode==0.1",
    "cryptography==44.0.1",
    "currency-symbols==2.0.3",
    "google-api-python-client==1.7.11",
    "google-auth-oauthlib==0.4.0",
    "herald-client",
    "insider-client",
    "json-excel-converter[xlsxwriter]==1.3.0",
    "katara-client",
    "kombu==5.5.4",
    "metroculus-client",
    "mysql-connector-python==9.1.0",
    "netaddr==0.7.19",
    "opentelemetry-api==1.29.0",
    "opentelemetry-sdk==1.29.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.29.0",
    "opentelemetry-instrumentation-logging==0.50b0",
    "optscale-data",
    "optscale-exceptions",
    "optscale-password",
    "optscale-time",
    "optscale-types",
    "protobuf>=5.29.1,<6",
    "pydevd-pycharm==201.5616.27",
    "pymongo==4.6.3",
    "pyyaml==6.0.1",
    "sqlalchemy==1.3.24",
    "tornado==6.5",
    "xlsxwriter==1.4.4",
]

[dependency-groups]
dev = [
    "check-alembic-down-revisions",
    "coverage==4.2",
    "freezegun==0.3.8",
    "mongomock==4.1.2",
    "pycodestyle==2.11.1",
    "pylint==3.0.2",
    "pytest==8.1.0",
    "pytest-xdist==3.5.0",
    "restapi-client",
    "setuptools==80.9.0",
]

[build-system]
requires = ["uv_build==0.8.15"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ".."
module-name = "rest_api"

[tool.uv.sources]
optscale-data = { path = "../tools/optscale_data", editable = true }
optscale-exceptions = { path = "../tools/optscale_exceptions", editable = true }
optscale-types = { path = "../tools/optscale_types", editable = true }
optscale-time = { path = "../tools/optscale_time", editable = true }
optscale-password = { path = "../tools/optscale_password", editable = true }
config-client = { path = "../optscale_client/config_client", editable = true }
auth-client = { path = "../optscale_client/auth_client", editable = true }
herald-client = { path = "../optscale_client/herald_client", editable = true }
katara-client = { path = "../optscale_client/katara_client", editable = true }
metroculus-client = { path = "../optscale_client/metroculus_client", editable = true }
insider-client = { path = "../optscale_client/insider_client", editable = true }
cloud-adapter = { path = "../tools/cloud_adapter", editable = true }
check-alembic-down-revisions = { path = "../tools/check_alembic_down_revisions", editable = true }
restapi-client = { path = "../optscale_client/rest_api_client", editable = true }
